requirements.txt
langchain
langchain-openai
python-dotenv
pydantic
langchain-google-genai
google-generativeai
pandas
matplotlib
scikit-learn
sentence-transformers
pymilvus